Import MSDRG weights

Imports a dataset containing MS-DRG's, descriptions, associated payment weights, and length of stay for FY2022 CMS Final Rule

cms_msdrg()

Value

A tibble::tibble

Author

Alex Zajichek

Source

https://www.cms.gov/medicare/acute-inpatient-pps/fy-2022-ipps-final-rule-home-page

Examples

cms_msdrg()
#> Rows: 767 Columns: 9
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (6): MS-DRG, FY 2022 Final Post-Acute DRG, FY 2022 Final Special Pay DRG...
#> dbl (3): Weights, Geometric mean LOS, Arithmetic mean LOS
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 767 × 7
#>    MSDRGCode MSDRGDescription      MSDRGType MajorDiagnostic… Weight GMLOS AMLOS
#>    <chr>     <chr>                 <chr>     <chr>             <dbl> <dbl> <dbl>
#>  1 001       HEART TRANSPLANT OR … SURG      PRE               28.9   30.1  39.1
#>  2 002       HEART TRANSPLANT OR … SURG      PRE               15.0   15.4  18.2
#>  3 003       ECMO OR TRACHEOSTOMY… SURG      PRE               19.1   22.4  30.2
#>  4 004       TRACHEOSTOMY WITH MV… SURG      PRE               11.9   20    24.6
#>  5 005       LIVER TRANSPLANT WIT… SURG      PRE               10.2   14.4  19.4
#>  6 006       LIVER TRANSPLANT WIT… SURG      PRE                4.70   7.5   8.1
#>  7 007       LUNG TRANSPLANT       SURG      PRE               11.6   17.4  20.8
#>  8 008       SIMULTANEOUS PANCREA… SURG      PRE                5.43   9    10.2
#>  9 010       PANCREAS TRANSPLANT   SURG      PRE                3.62   8     9.1
#> 10 011       TRACHEOSTOMY FOR FAC… SURG      PRE                5.02  10.9  13.8
#> # … with 757 more rows