Import a dataset from the Provider Data Catalog

Imports a specified dataset from the CMS Provider Data Catalog using readr::read_csv.

read_cms(
  datasetid = NULL,
  downloadurl = NULL,
  ...
)

Arguments

datasetid

A dataset identifier. Used only if the download URL is not specified.

downloadurl

The direct URL to the dataset of interest.

...

Additional arguments passed to readr::read_csv.

Value

A tibble::tibble

Author

Alex Zajichek

Examples

read_cms(
  datasetid = "qqw3-t4ie"
)
#> Rows: 19 Columns: 10
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (5): Measure ID, Measure Name, Number of Hospitals Too Few, Start Date, ...
#> dbl (4): National Rate, Number of Hospitals Worse, Number of Hospitals Same,...
#> lgl (1): Footnote
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 19 × 10
#>    `Measure ID` `Measure Name` `National Rate` `Number of Hos…` `Number of Hos…`
#>    <chr>        <chr>                    <dbl>            <dbl>            <dbl>
#>  1 COMP_HIP_KN… Rate of compl…            2.4                36             2668
#>  2 MORT_30_AMI  Death rate fo…           12.3                12             2120
#>  3 MORT_30_CABG Death rate fo…            2.9                17              921
#>  4 MORT_30_COPD Death rate fo…            8.1                53             3352
#>  5 MORT_30_HF   Death rate fo…           11.2               126             3074
#>  6 MORT_30_PN   Death rate fo…           15.3               221             3600
#>  7 MORT_30_STK  Death rate fo…           13.5                89             2168
#>  8 PSI_03       Pressure ulce…            0.59              200             3065
#>  9 PSI_04       Death rate am…          159.                 34             1489
#> 10 PSI_06       Iatrogenic pn…            0.23               13             3319
#> 11 PSI_08       In-hospital f…            0.1                 0             3326
#> 12 PSI_09       Perioperative…            2.55               25             2999
#> 13 PSI_10       Postoperative…            1.42               16             2753
#> 14 PSI_11       Postoperative…            5.03               50             2677
#> 15 PSI_12       Perioperative…            3.63               90             2922
#> 16 PSI_13       Postoperative…            4.9                36             2678
#> 17 PSI_14       Postoperative…            0.86                1             2691
#> 18 PSI_15       Abdominopelvi…            1.2                39             2949
#> 19 PSI_90       CMS Medicare …            1                 148             3180
#> # … with 5 more variables: `Number of Hospitals Better` <dbl>,
#> #   `Number of Hospitals Too Few` <chr>, Footnote <lgl>, `Start Date` <chr>,
#> #   `End Date` <chr>